requirements.txt
# Django framework
Django>=4.0,<5.0

# Django REST framework for building APIs
djangorestframework>=3.12.0,<4.0

# JWT authentication for DRF
djangorestframework-simplejwt>=5.0.0,<6.0

# PostgreSQL database adapter
psycopg2-binary>=2.8.6,<3.0

# CORS headers for allowing cross-origin requests (optional)
django-cors-headers>=3.7.0,<4.0

# For handling environment variables
python-dotenv>=0.19.0,<1.0

# To auto-generate API documentation
drf-spectacular>=0.22.0,<1.0
django-debug-toolbar>=3.2.0,<4.0
